+++ src/library/DIPS.Mobile.UI/Components/Alerting/Dialog/Android/DialogService.cs
@@ -1,10 +1,8 @@
-using Android.App;
 using DIPS.Mobile.UI.API.Library;
 using DIPS.Mobile.UI.Components.Alerting.Dialog.Android;
 using DIPS.Mobile.UI.Internal.Logging;
 using Java.Lang;
 using Microsoft.Maui.Platform;
-using AlertDialog = DIPS.Mobile.UI.Components.Alerting.Dialog.Android.AlertDialog;
 using Exception = System.Exception;
 
 // ReSharper disable once CheckNamespace
@@ -12,9 +10,7 @@ namespace DIPS.Mobile.UI.Components.Alerting.Dialog;
 
 public static partial class DialogService
 {
-    private const string AlertDialogTag = "MessageAlertTag";
-
-    private const string InputDialogTag = "InputDialogTag";
+    private const string DialogTag = "DUI.DialogFragment";
 
     public static partial async Task<InputDialogAction> ShowInputDialog(
         Action<IInputDialogConfigurator> configurator)
@@ -28,23 +24,18 @@ public static partial class DialogService
             
             IInputDialog inputDialogConfig = inputDialogConfigurator;
 
-            var inputDialog = new InputDialog(
+            var inputDialog = new DialogFragment(
                 inputDialogConfigurator,
-                "Ok",
-                "Cancel",
-                () => taskCompletionSource.TrySetResult(new InputDialogAction
+                () =>
                 {
-                    DialogAction = DialogAction.TappedAction,
-                    DialogInputs = inputDialogConfig.InputDialogEntryConfigurators 
-                }),
-                () => taskCompletionSource.TrySetResult(new InputDialogAction
+                    OnActionButtonTapped(inputDialogConfig, taskCompletionSource);
+                },
+                () =>
                 {
-                    DialogAction = DialogAction.Closed,
-                    DialogInputs = [] //TODO: Returner liste med tomme verdier
-                }));
+                    OnCancelButtonTapped(inputDialogConfig, taskCompletionSource);
+                });
              
-            inputDialog.Show(DUI.GetCurrentMauiContext?.Context?.GetFragmentManager()!, InputDialogTag);
-
+            inputDialog.Show(DUI.GetCurrentMauiContext?.Context?.GetFragmentManager()!, DialogTag);
         }
         catch(IllegalStateException)
         {
@@ -56,7 +47,7 @@ public static partial class DialogService
         }
         catch (Exception e)
         {
-            DUILogService.LogError<AlertDialog>(e.Message);
+            DUILogService.LogError<DialogFragment>(e.Message);
             taskCompletionSource.TrySetResult(new InputDialogAction
             {
                 DialogAction = DialogAction.Closed,
@@ -67,22 +58,37 @@ public static partial class DialogService
 
         return await taskCompletionSource.Task;
     }
-    
-    public static partial Task<DialogAction> ShowMessage(string title, string message, string actionTitle)
-    {
-        return Show(title, message, actionTitle);
-    }
 
-    public static partial Task<DialogAction> ShowConfirmationMessage(string title, string message, string closeTitle,
-        string actionTitle)
+    public static partial Task<DialogAction> ShowMessage(Action<IDialogConfigurator> configurator)
     {
-        return Show(title, message, actionTitle, closeTitle);
-    }
+        var dialogConfigurator = new DialogConfigurator();
+        configurator.Invoke(dialogConfigurator);
+        
+        IDialog dialog = dialogConfigurator;
+        
+        var taskCompletionSource = new TaskCompletionSource<DialogAction>();
 
-    public static partial Task<DialogAction> ShowDestructiveConfirmationMessage(string title, string message,
-        string closeTitle, string actionTitle)
-    {
-        return Show(title, message, actionTitle, closeTitle, true);
+        try
+        {
+            RemovePreviousDialog();
+
+            var alertDialog = new DialogFragment(dialog,
+                () => taskCompletionSource.TrySetResult(DialogAction.TappedAction),
+                () => taskCompletionSource.TrySetResult(DialogAction.Closed));
+            alertDialog.Show(DUI.GetCurrentMauiContext!.Context!.GetFragmentManager()!, DialogTag);
+        }
+        catch (IllegalStateException)
+        {
+            taskCompletionSource.TrySetResult(DialogAction.Closed);
+        }
+        catch (Exception e)
+        {
+            DUILogService.LogError<DialogFragment>(e.Message);
+            taskCompletionSource.TrySetResult(DialogAction.Closed);
+            throw;
+        }
+
+        return taskCompletionSource.Task;
     }
 
     public static partial Task Remove()
@@ -97,7 +103,7 @@ public static partial class DialogService
         }
         catch (Exception e)
         {
-            DUILogService.LogError<AlertDialog>(e.Message);
+            DUILogService.LogError<DialogFragment>(e.Message);
             throw;
         }
 
@@ -110,21 +116,21 @@ public static partial class DialogService
 
     }
 
-    internal static bool TryGetAlertDialog(out AlertDialog? alertDialog)
+    internal static bool TryGetAlertDialog(out DialogFragment? alertDialog)
     {
         var fragmentManager = DUI.GetCurrentMauiContext!.Context!.GetFragmentManager();
-        var previous = fragmentManager!.FindFragmentByTag(AlertDialogTag);
+        var previous = fragmentManager!.FindFragmentByTag(DialogTag);
         alertDialog = null;
-        if (previous is not AlertDialog theDialog)
+        if (previous is not DialogFragment theDialog)
         {
             return false;
         }
 
         alertDialog = theDialog;
         return true;
-
     }
 
+    [Obsolete("Remove when no references")]
     private static Task<DialogAction> Show(
         string title,
         string message,
@@ -132,40 +138,23 @@ public static partial class DialogService
         string? cancelButtonTitle = null,
         bool isDestructiveDialog = false)
     {
-        var taskCompletionSource = new TaskCompletionSource<DialogAction>();
-
-        try
+        return ShowMessage(config =>
         {
-            RemovePreviousDialog();
-
-            var alertDialog = new AlertDialog(title,
-                message,
-                actionButtonTitle,
-                cancelButtonTitle,
-                () => taskCompletionSource.TrySetResult(DialogAction.TappedAction),
-                () => taskCompletionSource.TrySetResult(DialogAction.Closed),
-                isDestructiveDialog);
-            alertDialog.Show(DUI.GetCurrentMauiContext!.Context!.GetFragmentManager()!, AlertDialogTag);
-        }
-        catch (IllegalStateException)
-        {
-            taskCompletionSource.TrySetResult(DialogAction.Closed);
-        }
-        catch (Exception e)
-        {
-            DUILogService.LogError<AlertDialog>(e.Message);
-            taskCompletionSource.TrySetResult(DialogAction.Closed);
-            throw;
-        }
-
-        return taskCompletionSource.Task;
+            config.SetTitle(title);
+            config.SetDescription(message);
+            config.SetActionTitle(actionButtonTitle);
+            if(cancelButtonTitle is not null)
+                config.SetCancelTitle(cancelButtonTitle);
+            if(isDestructiveDialog)
+                config.SetDestructive();
+        });
     }
 
     private static void RemovePreviousDialog()
     {
         var fragmentManager = DUI.GetCurrentMauiContext!.Context!.GetFragmentManager();
-        var previous = fragmentManager!.FindFragmentByTag(AlertDialogTag);
-        if (previous is AlertDialog alertDialog)
+        var previous = fragmentManager!.FindFragmentByTag(DialogTag);
+        if (previous is DialogFragment alertDialog)
         {
             alertDialog.Dismiss();
         }