+++ src/library/DIPS.Mobile.UI/Components/Alerting/Dialog/DialogService.cs
@@ -3,44 +3,6 @@ namespace DIPS.Mobile.UI.Components.Alerting.Dialog;
 public static partial class DialogService
 {
     public static partial Task<InputDialogAction> ShowInputDialog(Action<IInputDialogConfigurator> configurator);
-    
-    /// <summary>
-    ///     Attempt to show a message with a suggestion and one button.
-    /// </summary>
-    /// <param name="title">Title of the alert</param>
-    /// <param name="message">Message shown in alert</param>
-    /// <param name="actionTitle">Title of the button</param>
-    /// <returns>
-    ///     <see cref="DialogAction.TappedAction" />
-    ///     <see cref="DialogAction.Closed" />
-    /// </returns>
-    public static partial Task<DialogAction> ShowMessage(string title, string message, string actionTitle);
-
-    /// <summary>
-    ///     Show a message with a suggestion and two buttons.
-    /// </summary>
-    /// <param name="title">Title of the alert</param>
-    /// <param name="message">Message shown in alert</param>
-    /// <param name="closeTitle">Title of the cancel button</param>
-    /// <param name="actionTitle">Title of the action button</param>
-    /// <returns>
-    ///     <see cref="DialogAction.TappedAction" />
-    ///     <see cref="DialogAction.Closed" />
-    /// </returns>
-    public static partial Task<DialogAction> ShowConfirmationMessage(string title, string message, string closeTitle, string actionTitle);
-
-    /// <summary>
-    ///     Show a message with a suggestion and two buttons where the action button is a destructive button.
-    /// </summary>
-    /// <param name="title">Title of the alert</param>
-    /// <param name="message">Message shown in alert</param>
-    /// <param name="closeTitle">Title of the cancel button</param>
-    /// <param name="actionTitle">Title of the action button</param>
-    /// <returns>
-    ///     <see cref="DialogAction.TappedAction" />
-    ///     <see cref="DialogAction.Closed" />
-    /// </returns>
-    public static partial Task<DialogAction> ShowDestructiveConfirmationMessage(string title, string message, string closeTitle, string actionTitle);
 
     /// <summary>
     ///     Removes the currently showing alert.
@@ -52,8 +14,10 @@ public static partial class DialogService
     /// </summary>
     /// <returns>true/false depending on if its showing.</returns>
     public static partial bool IsShowing();
+
+    public static partial Task<DialogAction> ShowMessage(Action<IDialogConfigurator> configurator);
     
-    internal static bool ActionEnabledState(IInputDialog inputDialog)
+    internal static bool ActionEnabledState(IInputDialog? inputDialog)
     {
         return inputDialog.InputDialogEntryConfigurators.All(s =>
         {
@@ -65,12 +29,35 @@ public static partial class DialogService
             return true;
         });
     }
+    
+    private static void OnCancelButtonTapped(IInputDialog inputDialogConfig, TaskCompletionSource<InputDialogAction> taskCompletionSource)
+    {
+        inputDialogConfig.InputDialogEntryConfigurators.ForEach(input =>
+        {
+            input.Reset();
+        });
+
+        taskCompletionSource.TrySetResult(new InputDialogAction
+        {
+            DialogAction = DialogAction.Closed,
+            DialogInputs = inputDialogConfig.InputDialogEntryConfigurators
+        });
+    }
+    
+    private static void OnActionButtonTapped(IInputDialog inputDialogConfig, TaskCompletionSource<InputDialogAction> taskCompletionSource)
+    {
+        taskCompletionSource.TrySetResult(new InputDialogAction
+        {
+            DialogAction = DialogAction.TappedAction,
+            DialogInputs = inputDialogConfig.InputDialogEntryConfigurators
+        });
+    }
 }
 
 public class InputDialogAction
 {
     public DialogAction DialogAction { get; set; }
-    public IEnumerable<IDialogInputField> DialogInputs { get; set; }
+    public IEnumerable<IDialogInputField> DialogInputs { get; set; } = [];
 }