+++ src/library/DIPS.Mobile.UI/Components/Alerting/Dialog/InputDialogConfigurator.cs
@@ -1,6 +1,6 @@
 namespace DIPS.Mobile.UI.Components.Alerting.Dialog;
 
-public class InputDialogConfigurator : IInputDialogConfigurator, IInputDialog
+public class InputDialogConfigurator : DialogConfigurator, IInputDialogConfigurator, IInputDialog
 {
     public IInputDialogConfigurator AddInputField<T>(IDialogInputField<T> dialogInputField)
     {
@@ -11,7 +11,7 @@ public class InputDialogConfigurator : IInputDialogConfigurator, IInputDialog
     List<IDialogInputField> IInputDialog.InputDialogEntryConfigurators { get; } = [];
 }
     
-public interface IInputDialogConfigurator
+public interface IInputDialogConfigurator : IDialogConfigurator
 {
     IInputDialogConfigurator AddInputField<T>(IDialogInputField<T> inputField);
 }
@@ -19,25 +19,39 @@ public interface IInputDialogConfigurator
 public interface IDialogInputField
 {
     internal Guid Identifier { get; }
+    
+    /// <summary>
+    /// Resets the value of the input field to the value it had when it was created.
+    /// </summary>
+    void Reset();
 }
 
 public interface IDialogInputField<T> : IDialogInputField
 {
-    public T Value { get; }
-    public bool MustBeSet { get; set; }
+    T Value { get; }
+    bool MustBeSet { get; set; }
 }
 
 public class StringDialogInputField : IDialogInputField<string>
 {
+    private readonly string m_valueOnCreation;
+
     public StringDialogInputField(string? placeholder = null, string? value = null, bool mustBeSet = false)
     {
         Placeholder = placeholder ?? string.Empty;
         Value = value ?? string.Empty;
         MustBeSet = mustBeSet;
+        m_valueOnCreation = Value;
     }
     public string Placeholder { get; set; }
     public string Value { get; set; }
     public bool MustBeSet { get; set; }
+
+    public void Reset()
+    {
+        Value = m_valueOnCreation;
+    }
+
     public Guid Identifier { get; } = Guid.NewGuid();
 }