+++ src/library/DIPS.Mobile.UI/Components/Alerting/Dialog/iOS/DialogService.cs
@@ -1,4 +1,3 @@
-using DIPS.Mobile.UI.Resources.LocalizedStrings.LocalizedStrings;
 using Microsoft.Maui.Platform;
 using UIKit;
 using Colors = Microsoft.Maui.Graphics.Colors;
@@ -8,48 +7,30 @@ namespace DIPS.Mobile.UI.Components.Alerting.Dialog;
 
 public static partial class DialogService
 {
-    private static TaskCompletionSource<DialogAction>? m_taskCompletionSource;
-    private static TaskCompletionSource<InputDialogAction>? m_inputDialogCompletionSource;
+    private static TaskCompletionSource<InputDialogAction>? s_inputTaskCompletionSource;
+    private static TaskCompletionSource<DialogAction>? s_taskCompletionSource;
     private static UIWindow Window { get; } = new() { BackgroundColor = Colors.Transparent.ToPlatform() };
 
     public static partial async Task<InputDialogAction> ShowInputDialog(Action<IInputDialogConfigurator> configurator)
     {
         await Remove();
-        m_inputDialogCompletionSource = new TaskCompletionSource<InputDialogAction>();
+        
+        s_inputTaskCompletionSource = new TaskCompletionSource<InputDialogAction>();
         var inputDialogConfigurator = new InputDialogConfigurator();
         
         configurator.Invoke(inputDialogConfigurator);
 
+        IDialog dialog = inputDialogConfigurator;
         IInputDialog inputDialog = inputDialogConfigurator;
-        
-        var uiAlertController = UIAlertController.Create("Alert", string.Empty, UIAlertControllerStyle.Alert);
-        
-        uiAlertController.AddAction(
-            UIAlertAction.Create(
-                "Cancel",
-                UIAlertActionStyle.Cancel,
-                _ =>
-                {
-                    Window.Hidden = true;
-                    m_inputDialogCompletionSource?.SetResult(new InputDialogAction
-                    {
-                        DialogAction = DialogAction.Closed,
-                        DialogInputs = [] //TODO: Returner liste med tomme verdier
-                    });
-                }));
 
-        var okAction = UIAlertAction.Create(
-            "Ok",
-            UIAlertActionStyle.Default,
-            _ =>
-            {
-                Window.Hidden = true;
-                m_inputDialogCompletionSource?.SetResult(new InputDialogAction
-                {
-                    DialogAction = DialogAction.TappedAction,
-                    DialogInputs = inputDialog.InputDialogEntryConfigurators
-                });
-            });
+        var uiAlertController = ConstructGeneralAlertController(dialog, () =>
+        {
+            OnCancelButtonTapped(inputDialogConfigurator, s_inputTaskCompletionSource);
+        },
+        () =>
+        {
+            OnActionButtonTapped(inputDialog, s_inputTaskCompletionSource);
+        });
         
         foreach (var dialogInputField in inputDialog.InputDialogEntryConfigurators)
         {
@@ -60,58 +41,68 @@ public static partial class DialogService
                     e.RestorationIdentifier = stringInputField.Identifier.ToString();
                     e.Placeholder = stringInputField.Placeholder;
                     e.Text = stringInputField.Value;
-                    e.EditingChanged += (sender, args) =>
+                    e.EditingChanged += delegate
                     {
                         stringInputField.Value = e.Text;
-                        okAction.Enabled = ActionEnabledState(inputDialog);
+                        
+                        SetActionButtonEnabled();
                     };
                 });
             }
         }
-
-        okAction.Enabled = ActionEnabledState(inputDialog);
-
-        uiAlertController.AddAction(okAction);
         
-        Window.RootViewController = new UIViewController();
-        if (Window.RootViewController.View == null)
+        void SetActionButtonEnabled()
         {
-            return await m_inputDialogCompletionSource.Task;
+            var actionButton = uiAlertController.Actions.FirstOrDefault(action => action.Title == dialog.ActionTitle);
+            if (actionButton is not null)
+            {
+                actionButton.Enabled = ActionEnabledState(inputDialog);
+            }
         }
-        
-        Window.RootViewController.View.BackgroundColor = Colors.Transparent.ToPlatform();
-        Window.WindowLevel = UIWindowLevel.Alert + 1;
-        Window.MakeKeyAndVisible();
-        Window.RootViewController.PresentViewController(uiAlertController, true, () => { });
-        
 
-        return await m_inputDialogCompletionSource.Task;
+        SetActionButtonEnabled();
+        
+        Window.RootViewController?.PresentViewController(uiAlertController, true, () => { });
+        
+        return await s_inputTaskCompletionSource.Task;
+    }
+
+    public static partial async Task<DialogAction> ShowMessage(Action<IDialogConfigurator> configurator)
+    {
+        var dialogConfigurator = new DialogConfigurator();
+        configurator.Invoke(dialogConfigurator);
+        
+        IDialog dialog = dialogConfigurator;
+        
+        await Remove();
+        
+        s_taskCompletionSource = new TaskCompletionSource<DialogAction>();
+        var uiAlertController = ConstructGeneralAlertController(dialog, () =>
+            {
+                s_taskCompletionSource.SetResult(DialogAction.Closed);
+            },
+            () =>
+            {
+                s_taskCompletionSource.SetResult(DialogAction.TappedAction);
+            }, dialog.IsDestructive);
+       
+        Window.RootViewController?.PresentViewController(uiAlertController, true, () => { });
+
+        return await s_taskCompletionSource.Task;
     }
     
-    public static partial Task<DialogAction> ShowMessage(string title, string message, string actionTitle)
-    {
-        return Show(title, message, actionTitle);
-    }
-
-    public static partial Task<DialogAction> ShowConfirmationMessage(string title, string message, string closeTitle,
-        string actionTitle)
-    {
-        return Show(title, message, actionTitle, closeTitle);
-    }
-
-    public static partial Task<DialogAction> ShowDestructiveConfirmationMessage(string title, string message,
-        string closeTitle, string actionTitle)
-    {
-        return Show(title, message, actionTitle, closeTitle, true);
-    }
-
     public async static partial Task Remove()
     {
         if (TryGetUiAlertViewController(out var viewController))
         {
             await viewController?.DismissViewControllerAsync(false)!;
             Window.Hidden = true;
-            m_taskCompletionSource?.TrySetResult(DialogAction.Closed);   
+            s_taskCompletionSource?.TrySetResult(DialogAction.Closed);  
+            s_inputTaskCompletionSource?.TrySetResult(new InputDialogAction
+            {
+                DialogAction = DialogAction.Closed,
+                DialogInputs = []
+            });
         }
     }
 
@@ -125,7 +116,6 @@ public static partial class DialogService
 
         viewController = Window.RootViewController;
         return true;
-
     }
     
     public static partial bool IsShowing()
@@ -133,52 +123,38 @@ public static partial class DialogService
         return TryGetUiAlertViewController(out _);
     }
 
-    private async static Task<DialogAction> Show(
-        string title,
-        string message,
-        string actionButtonTitle,
-        string? cancelButtonTitle = null,
-        bool isDestructiveDialog = false)
+    private static UIAlertController ConstructGeneralAlertController(IDialog dialog, Action onCancelButtonTapped, Action onActionButtonTapped, bool isDestructiveDialog = false)
     {
-        await Remove();
-
-        m_taskCompletionSource = new TaskCompletionSource<DialogAction>();
-
-        var uiAlertController = UIAlertController.Create(title, message, UIAlertControllerStyle.Alert);
-        if (cancelButtonTitle is not null)
+        var uiAlertController = UIAlertController.Create(dialog.Title, dialog.Description, UIAlertControllerStyle.Alert);
+        if (dialog.CancelTitle is not null)
         {
             uiAlertController.AddAction(
                 UIAlertAction.Create(
-                    cancelButtonTitle,
+                    dialog.CancelTitle,
                     UIAlertActionStyle.Default,
                     _ =>
                     {
                         Window.Hidden = true;
-                        m_taskCompletionSource?.SetResult(DialogAction.Closed);
+                        onCancelButtonTapped.Invoke();
                     }));
         }
 
         uiAlertController.AddAction(
             UIAlertAction.Create(
-                actionButtonTitle,
+                dialog.ActionTitle,
                 isDestructiveDialog ? UIAlertActionStyle.Destructive : UIAlertActionStyle.Default,
                 _ =>
                 {
                     Window.Hidden = true;
-                    m_taskCompletionSource?.SetResult(DialogAction.TappedAction);
+                    onActionButtonTapped.Invoke();
                 }));
         
         Window.RootViewController = new UIViewController();
-        if (Window.RootViewController.View == null)
-        {
-            return await m_taskCompletionSource.Task;
-        }
         
-        Window.RootViewController.View.BackgroundColor = Colors.Transparent.ToPlatform();
+        Window.RootViewController.View!.BackgroundColor = Colors.Transparent.ToPlatform();
         Window.WindowLevel = UIWindowLevel.Alert + 1;
         Window.MakeKeyAndVisible();
-        Window.RootViewController.PresentViewController(uiAlertController, true, () => { });
 
-        return await m_taskCompletionSource.Task;
+        return uiAlertController;
     }
 }